+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 
 ## New Features
+- #1471 Unbounded partitioned windows 
 - #1445 Support for CURRENT_DATE, CURRENT_TIME and CURRENT_TIMESTAMP
+- #1505 Support for right outer join
 
 
 ## Improvements